Do more rust-from-c

+++ rust-from-c/src/lib.rs
@@ -1,5 +1,5 @@
 #[no_mangle]
-pub extern "C" fn inc(x: usize) -> usize {
+pub extern "C" fn inc(x: i32) -> i32 {
   println!("hello from rust inc({})", x);
   x + 1
 }
